Update graph prototype with db connection

+++ db/management.py
@@ -8,6 +8,9 @@ import sqlite3
 import csv
 import os
 
+inp = input('\nAre you sure to delete data.db ? y/n: ')
+if inp == 'n':
+    exit()
 os.system('rm -rf data.db && touch data.db')
 conn = sqlite3.connect('data.db')
 cur = conn.cursor()
@@ -17,6 +20,7 @@ cur.execute(
 """
 CREATE TABLE data(
     timestamp VARCHAR(40),
+    timeinsec VARCHAR(40),
     bmp280 VARCHAR(40),
     hdc1080 VARCHAR(40),
     neo6m VARCHAR(40),
@@ -35,7 +39,8 @@ print('[ ok ] Db has been sucesfully created!')
 db = SQLAlchemy()
 
 class Data(db.Model):
-    timestamp = db.Column(db.String(40), primary_key=True, unique=True)
+    timestamp = db.Column(db.String(40), primary_key=True, unique=True) 
+    timeinsec = db.Column(db.String(40))
     bmp280 = db.Column(db.String(40))
     hdc1080 = db.Column(db.String(40))
     neo6m = db.Column(db.String(40))
@@ -48,9 +53,10 @@ Data_tbl = Table('data', Data.metadata)
 def create_data_table():
     Data.metadata.create_all(engine)
 
-def add_data(timestamp, bmp280, hdc1080, neo6m, mpu9250, camera):
+def add_data(timestamp, timeinsec, bmp280, hdc1080, neo6m, mpu9250, camera):
     ins = Data_tbl.insert().values(
         timestamp=timestamp,
+        timeinsec=timeinsec,
         bmp280=bmp280,
         hdc1080=hdc1080,
         neo6m=neo6m,
@@ -69,7 +75,7 @@ def del_data(code):
 
 
 def show_data():
-    select_st = select([Data_tbl.c.timestamp, Data_tbl.c.bmp280, Data_tbl.c.hdc1080, Data_tbl.c.neo6m, Data_tbl.c.mpu9250, Data_tbl.c.camera])
+    select_st = select([Data_tbl.c.timestamp, Data_tbl.c.timeinsec, Data_tbl.c.bmp280, Data_tbl.c.hdc1080, Data_tbl.c.neo6m, Data_tbl.c.mpu9250, Data_tbl.c.camera])
     conn = engine.connect()
     rs = conn.execute(select_st)
     for row in rs:
